requirements.txt
fastapi
requests
playwright
beautifulsoup4
aiohttp
sqlalchemy
python-dotenv
